LPS-80570 Move APIO Form helpers to util package
